requirements.txt
dash==2.11.0
dash_bootstrap_components==1.4.1
ipython==8.14.0
pandas==2.0.2
Pillow==9.5.0
plotly==5.15.0
gunicorn
dash-tools
